+++ packages/quantic/force-app/main/default/lwc/quanticQueryError/quanticQueryError.js
@@ -49,14 +49,13 @@ export default class QuanticQueryError extends LightningElement {
   }
 
   renderedCallback() {
-    initializeWithHeadless(this, this.engineId, this.initialize.bind(this));
+    initializeWithHeadless(this, this.engineId, this.initialize);
   }
 
   /**
    * @param {import("coveo").SearchEngine} engine
    */
-  @api
-  initialize(engine) {
+  initialize = (engine) => {
     this.queryError = CoveoHeadless.buildQueryError(engine);
     this.unsubscribe = this.queryError.subscribe(() => this.updateState());
   }